### 2025-04-06 #7999 @editorialbot
[PRE REVIEW]: mlgidGUI - an annotation program for 2D scattering data with powder geometry
**Submitting author:** @cvoelt (<a href="http://orcid.org/0009-0001-2690-4468">Constantin Völter</a>)
**Repository:** https://github.com/mlgid-project/mlgidGUI
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.8.6
