+++ docs/HEADER.yaml
@@ -1,6 +1,5 @@
 ---
  title: "mlgidGUI"
- author: "Authors: Constantin Völter and Vladimir Starostin"
  date:
  subtitle: "User Manual"
  geometry: "left=2.54cm,right=2.54cm,top=1.91cm,bottom=1.91cm"
