### 2025-05-12 #7999 @lucydot
@novak-jiri @kowarik-lab @stuartcampbell @jwuttke - hello 👋  - are you available to review this submission?

If you have not reviewed for JOSS, welcome! We are a platinum open access (free-to-author, free-to-read) , "developer-friendly" journal for open source research software. Our review process is fully transparent - it is a discussion between authors, reviewers and editors, with the aim of increasing the quality of the submitted software. For more details, please see [our docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). You can also ask me questions here.
