### 2025-06-27 #8499 @editorialbot
[REVIEW]: mlgidGUI - an annotation program for 2D scattering data with powder geometry
**Submitting author:** @cvoelt (<a href="http://orcid.org/0009-0001-2690-4468">Constantin Völter</a>)
**Repository:** https://github.com/mlgid-project/mlgidGUI
**Branch with paper.md** (empty if default branch):
**Version:** 0.5.1
