### 2025-07-16 #8499 @lucydot
A quick message to let you know I'll be on leave (without access to internet) tomorrow until the 4th of August, when I will check back here.
