### 2025-08-14 mlgid-project/mlgidGUI#1 @kstenio
Paper improvements (JOSS Review)
I have nothing to point out about the manuscript. It is well written and gets straight to the point. However, I think it could benefit from some adjustments:

### 1st

I recommend that you add more figures of the app in the paper. A general figure showing the app and another showing its use, or an assembly showing several input files, or output files. It would make clearer to the reader where/how to use the app, as well as the pattern/response obtained after the analysis. Other suggestions:

   - Comparison between raw and processed GIWAXS images
   - Example of CIF-based peak simulation overlays
   - Workflow diagram showing data pipeline integration

### 2nd

There are 2 contributors in the repository, but 7 in the paper. This is normal, since not everyone who helped with conceptualization or even algorithm development is exactly a programmer. That said, a good way to properly describe what each person did in the work is using [Contributor Role Taxonomy (CRediT)](https://credit.niso.org/)). In general, you would add a new field, "Author contributions", before "Acknowledgments", and in it describe what each person did/focused/acted. This makes the work much more transparent, as well as being an opportunity to rethink the authors (maybe someone was missing, maybe someone should be removed, this type of reflection). [This link has an example](https://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement), but there are many other ways to do it.
