### 2025-08-14 mlgid-project/mlgidGUI#2 @kstenio
Paper generation artifacts should be in another action (JOSS Review)
You are using the same GitHub Action to generate application artifacts and the paper. It would be recommended to use a secondary action for this, so that new releases do not necessarily regenerate new versions of the PDF. I would have recommended a separate branch for the paper, but now that it has entered the review stage, I believe it is too late.
