### 2025-08-14 mlgid-project/mlgidGUI#3 @kstenio
Could be distributed as a Python Package (JOSS Review)
You are using the old format with the `setup.py` file to install the program via PIP. This behavior is no longer recommended, with the encouragement to use [pyptoject.toml](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/). It has been more complicated to distribute Python packages, but today with [uv](https://docs.astral.sh/uv/), it is much simpler. You can initialize the project with `uv init`, add dependencies with `uv add [deps]` and development dependencies with `uv add [devs] --dev`. Then it is possible to compile a good wheel file with `uv build`. I also recommend publishing to [PyPI](https://pypi.org/), so anyone could install your app simply with `pip install mlgidGUI` (which is very cool). In this case, it would be a matter of configuring an action and, in it, calling `uv publish`.

Still on the same domain, have you thought about using semantic versioning with [MAJOR.MINOR.PATCH](https://peps.python.org/pep-0440/)?

These are tips, if you don't want to, you don't need to follow them, but they are highly recommended.
