### 2025-08-14 mlgid-project/mlgidGUI#4 @kstenio
Python 3.8 is in EOL (end of life) (JOSS Review) [critical]
By default, your project already generates binary artifacts (which is great), both for Linux and for Windows. However, the stack is based on Python 3.8, which has already left the support cycle, including security bug fixes (https://devguide.python.org/versions/).

My personal recommendation would be to migrate to Python 3.12, which entered the security stage this year. Many things have changed from Python 3.8 to 3.12, including libraries, so this process can be a bit more painful. Still, I strongly recommend the migration. If not to 3.12, at least to **3.9** (although this will enter EOL still in 2025).

You could start by initializing the repo with `uv init --python 3.12` and see if you can install all the libs, and make the program behave appropriately. One good thing is that you have tests ready, so if the tests pass, it is already a good indication.
