### 2025-08-14 #8499 @kstenio
Hi @lucydot and @cvoelt,

Sorry for taking this long to complete my review.

My assessment is that the mlgidGUI paper and software meet JOSS standards for scientific software publication. It represents a valuable contribution to the scientific software ecosystem, addressing a genuine need in the GIWAXS analysis community. The software demonstrates evidence of being developed by domain experts with an understanding of both scientific requirements and practical usability needs.

I had some minor concerns, which I’ve added to the software repository as issues. Most of these are just recommendations:

 * mlgid-project/mlgidGUI#1
 * mlgid-project/mlgidGUI#2
 * mlgid-project/mlgidGUI#3
 * mlgid-project/mlgidGUI#4

That said, I would recommend this paper for publication.
