### 2025-08-14 a4d9f5f
Python 3.12 support
