+++ AppImageBuilder.yml
@@ -5,64 +5,54 @@ script:
   - rm -rf AppDir  | true
   # Make usr and icons dirs
   - mkdir -p AppDir/usr/src
-  - mkdir -p AppDir/usr/share/icons/default/256x256/
+  - mkdir -p AppDir/usr/share/icons/hicolor/256x256/apps/
   # Copy the python application code into the AppDir
   - cp ./main.py  AppDir/usr/src/
-  - cp ./mlgidGUI/static/icons/window_icon_256.png AppDir/usr/share/icons/default/256x256/
-  # Install application dependencies
-  - python3.8 -m pip install --ignore-installed --no-deps --prefix=/usr --root=AppDir qdarkgraystyle==1.0.2
-  - python3.8 -m pip install --ignore-installed --prefix=/usr --root=AppDir deprecated
-  # Remove PyQt5 and qdarkgraystyle dependencies of mlgidGUI
-  - sed -i "/\b\(PyQt5\|qdarkgraystyle\|PyQt5_sip\)\b/d" setup.py
+  - cp ./mlgidGUI/static/icons/window_icon_256.png AppDir/usr/share/icons/hicolor/256x256/apps/
+
   # Install mlgidGUI
-  - python3.8 -m pip install --ignore-installed --prefix=/usr --root=AppDir ./
+  - python3.12 -m pip install --ignore-installed --prefix=/usr --root=AppDir ./
 
 AppDir:
 
   path: ./AppDir
- 
+
   app_info:
     id: org.appimage-crafters.python-appimage-example
     name: mlgidGUI
     icon: window_icon_256
     version: 0.1.0
     # Set the python executable as entry point
-    exec: /usr/bin/python3.8
+    exec: /usr/bin/python3.12
     # Set the application main script path as argument. Use '$@' to forward CLI parameters
     exec_args: "-s $APPDIR/usr/src/main.py"
 
   apt:
     arch: amd64
     sources:
-      - sourceline: 'deb [arch=amd64] http://de.archive.ubuntu.com/ubuntu focal main restricted universe multiverse'
-        key_url: 'http://keyserver.ubuntu.com/pks/lookup?op=get&search=0x3b4fe6acc0b21f32'
+      - sourceline: 'deb [arch=amd64] http://archive.ubuntu.com/ubuntu jammy main restricted universe multiverse'
+        key_url: 'http://keyserver.ubuntu.com/pks/lookup?op=get&search=0x871920D1991BC93C'
+
+      - sourceline: 'deb [arch=amd64] http://ppa.launchpad.net/deadsnakes/ppa/ubuntu jammy main'
+        key_url: 'http://keyserver.ubuntu.com/pks/lookup?op=get&search=0x6A755776'
 
     include:
-      - python3.8
+      - python3.12
       - python3-pkg-resources
-      - python3-opengl
-      - python3.8-dev
-      - python3-numpy
-      - squashfs-tools
-      - gsettings-desktop-schemas
-      - libgtk-3-common
-      - python3-pip
-      - python3-pyqt5
-      - libxcb-xinerama0
-      
+
     exclude: []
-    
+
   runtime:
     env:
       # Set python home
       # See https://docs.python.org/3/using/cmdline.html#envvar-PYTHONHOME
       PYTHONHOME: '${APPDIR}/usr/'
-      PATH: '${APPDIR}/usr/lib/python3.8/site-packages:${APPDIR}/usr/bin:${PATH}'
+      PATH: '${APPDIR}/usr/lib/python3.12/site-packages:${APPDIR}/usr/bin:${PATH}'
       # Path to the site-packages dir or other modules dirs
[…]

[…]
