### 2025-08-14 mlgid-project/mlgidGUI#4 @cvoelt
Thank you for the feedback, we updated the project to Python 3.12 in the commit a4d9f5fbc9f47648706720f4120329d2f7b325e2
