### 2025-08-15 576b3b5
Added PyPI support, changed version numbering to  MAJOR.MINOR.PATCH
