+++ .github/workflows/build_releases.yml
@@ -6,17 +6,16 @@ on:
   pull_request:
     branches: [ "main" ]
 
-jobs:  
+jobs:
   create_release:
-  
     runs-on: ubuntu-22.04
-    
+
     outputs:
       upload_url: ${{steps.create_release.outputs.upload_url}}
       version: ${{ steps.set_variables.outputs.VERSION }}
-      
+
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: Set variables
         id: 'set_variables'
@@ -24,7 +23,7 @@ jobs:
           VERSION=$(grep '^version =' pyproject.toml | sed -E "s/version = \"(.*)\"/\1/")
           echo "VERSION=$VERSION" >> $GITHUB_OUTPUT
 
-      - name: Delete old release                
+      - name: Delete old release
         uses: dev-drprasad/delete-older-releases@v0.2.1
         with:
           keep_latest: 0
@@ -50,7 +49,7 @@ jobs:
     runs-on: ubuntu-22.04
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - run: sed -i 's+./docs/WORKFLOW.md+#workflow+g' README.md
       - run: sed -i 's+./docs/DOCUMENTATION.md+#documentation+g' README.md
       - run: cp -r docs/images ./
@@ -93,12 +92,79 @@ jobs:
           asset_name: paper.pdf
           asset_content_type: application/pdf
 
-  test_and_win_exe:
-    needs: create_release
-    runs-on: windows-latest
+
+  build:
+    name: Build wheel
+    runs-on: ubuntu-22.04
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
+
+      - name: Patch pyqtgraph dependency for PyPi
+        run: |
+          sed -i "s|pyqtgraph @ git+https://github.com/pyqtgraph/pyqtgraph.git@6eb0bf1a928c3e8fef332bbebe8a9da3be0ab19a|pyqtgraph >= 0.13.0|" pyproject.toml
+          cat pyproject.toml
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.12"
+
+      - name: Install build tools
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install build
+
+      - name: Build wheel
+        run: python -m build --wheel
+
+      - name: Upload wheel artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: wheels-${{ matrix.os }}
+          path: dist/*.whl
+
+  publish:
+    name: Publish to PyPI
+    needs: [build, test_and_win10_exe]
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.12"
+
+      - name: Install twine
+        run: python -m pip install twine
+
+      - name: Download all wheel artifacts
+        uses: actions/download-artifact@v4
+        with:
+          path: dist
+
+      - name: Publish to PyPI
+        env:
+          TWINE_USERNAME: __token__
+          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
+        run: |
+          set +e
+          twine upload --verbose --repository-url https://pypi.org/legacy/ dist/**/*.whl
+          EXIT_CODE=$?
+          if [ "$EXIT_CODE" -eq 1 ]; then
+            echo "Upload failed — possibly due to duplicate file. Skipping gracefully."
+            exit 0
+          fi
[…]
