### 2025-08-15 mlgid-project/mlgidGUI#3 @cvoelt
Thank you for your feedback. We added a pyproject.toml file in the commit a4d9f5fbc9f47648706720f4120329d2f7b325e2 . The PyPI support and the versioning with MAJOR.MINOR.PATCH is introduced in dcecb9bfc03817552f70a17101078390e21ded04
