### 2025-08-17 dcecb9b
Added PyPI support, changed version numbering to MAJOR.MINOR.PATCH
