+++ mlgidGUI/__version.py
@@ -1,1 +1,1 @@
-__version__ = '0.4.0.0'
+__version__ = '0.4.1'
