+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "mlgidGUI"
-version = "0.4.1"
+version = "0.5.0"
 description = "A GUI program for GIWAXS images analysis and annotation"
 readme = "README.md"
 requires-python = ">=3.12.0"
