### 2025-08-17 mlgid-project/mlgidGUI#2 @cvoelt
Thank you for the feedback. The paper generation is now a separate action and is manually triggered only. The implementation is done in commit 317704892c369c8f4d46f3a6524e1dc9739d562a
