### 2025-08-28 e7bff9c
Updated paper, added logo
