+++ docs/paper.md
@@ -44,15 +44,33 @@ It supports the conversion to polar coordinates based on experimental settings, 
 The export as PASCAL VOC dataset enables the integration to ML pipelines. A ML-based peak detection [@starostin2022endtoend] can provide preliminary peaks to accelerate the annotation.
 
 # Statement of need
+
 Grazing-incidence wide-angle X-ray scattering (GIWAXS) is a powerful technique for characterizing crystalline structures on surfaces. It is instrumental in a wide range of applications, including organic photovoltaics and semiconductors [@feidenhansl1989surface; @banerjee2020grazing]. Typical GIWAXS measurements at synchrotron facilities generate hundreds of thousands of diffraction images per day. While recent ML tools aid in analyzing these vast datasets, proper benchmarking on annotated datasets is crucial for ensuring robust scientific applications.
 
-In our recent paper [@volter2025benchmarking], we developed a methodology for benchmarking GIWAXS data and published the first annotated dataset. Here, we open-source the software used for annotating this data to further support efforts in improving, standardizing, and automating GIWAXS analysis. With mlgidGUI, we support researchers in producing GIWAXS datasets with less effort and higher-quality annotations.
+In our recent paper [@volter2025benchmarking], we developed a methodology for benchmarking GIWAXS data and published the first annotated dataset. Here, we open-source the software used for annotating this data to further support efforts in improving, standardizing, and automating GIWAXS analysis. With mlgidGUI, we support researchers in producing GIWAXS datasets with less effort and higher-quality annotations. Figure 1 shows the appearance of mlgidGUI:
 
-# General Workflow
+![Appearance of the mlgidGUI: Project Manager on the left, Angular Profile at the bottom, Image Viewer in the center, and Radial Fit widget on the right.](images%2Fgui-overview.png)
+
+# General workflow
+
 mlgidGUI is intended for GIWAXS images which are already converted to reciprocal space. Optionally, the converted image can be analyzed beforehand by an automated peak detection [@starostin2022tracking] to accelerate the annotation process. The contrast correction is enabled by default to aid in finding peaks with low intensities. Next, the user has to set the geometry settings to ensure the correct beam center position and Q-range. Then, diffraction peaks are added manually with optional preferred crystallographic orientation. Each peak can be fitted with a chosen profile and background to ensure accurate position, intensity, and width. According to the proposed methodology in [@volter2025benchmarking], each peak is assigned with a confidence label - low, medium, or high. Additionally, an integrated crystallographic toolkit can be used to superimpose the measured data with the simulated diffraction pattern of a provided structure. These simulated peaks help the user in identifying weak peaks. The intermediate result of the labeling process is saved automatically. When an image is fully labeled, the resulting annotation can be saved as a new HDF5 file, added to an existsing HDF5 dataset or exported as PASCAL VOC dataset. 
+Figure 2 outlines the entire workflow: Raw detector data is first converted and optionally undergoes peak detection before it is visualized and annotated with the mlgidGUI. The resulting datasets from mlgidGUI can be used for training or validation purposes of the machine learning model. 
 
-# Related Work
+![End-to- end workflow from raw detector output through data conversion followed by visualization and annotation for the ML-based peak detection](images%2Fworkflow.png)
+
+# Related work
[…]
