### 2025-08-28 mlgid-project/mlgidGUI#1 @cvoelt
Thank you for the feedback! We included updated the paper with two figures and added the CRediT contributions in commit e7bff9c593e994529ebb670778cd90fbe14b8fe5 .
