### 2025-08-28 #8499 @cvoelt
> Hi [@lucydot](https://github.com/lucydot) and [@cvoelt](https://github.com/cvoelt),
>
> Sorry for taking this long to complete my review.
>
> My assessment is that the mlgidGUI paper and software meet JOSS standards for scientific software publication. It represents a valuable contribution to the scientific software ecosystem, addressing a genuine need in the GIWAXS analysis community. The software demonstrates evidence of being developed by domain experts with an understanding of both scientific requirements and practical usability needs.
>
> I had some minor concerns, which I’ve added to the software repository as issues. Most of these are just recommendations:
>
>     * [Paper improvements (JOSS Review) mlgid-project/mlgidGUI#1](https://github.com/mlgid-project/mlgidGUI/issues/1)
>
>     * [Paper generation artifacts should be in another action (JOSS Review) mlgid-project/mlgidGUI#2](https://github.com/mlgid-project/mlgidGUI/issues/2)
>
>     * [Could be distributed as a Python Package (JOSS Review) mlgid-project/mlgidGUI#3](https://github.com/mlgid-project/mlgidGUI/issues/3)
>
>     * [Python 3.8 is in EOL (end of life) (JOSS Review) [critical] mlgid-project/mlgidGUI#4](https://github.com/mlgid-project/mlgidGUI/issues/4)
>
>
> That said, I would recommend this paper for publication.

Thank you @kstenio for the valuable feedback and comments. We carefully implemented the requested changes, the commits are mentioned in the corresponding issues.  We revised the paper and added two figures and the contributions each author made.  We remain very excited by the first 2D scattering data annotation program for  ML datasets and hope that with these revisions our work can soon be accepted for JOSS.
