### 2025-10-23 #8499 @lucydot
Hi @cvoelt -  I have one small question about the manuscript. What does the "ROR" in the affiliations list mean?
