### 2025-10-23 #8499 @cvoelt
Thank you @lucydot for moving to the final editorial checks. Regarding your question, the ror stands for Research Organization Registry (https://ror.org/). We uploaded the current release 0.5.1 to zenodo. It can be found under the following doi: https://doi.org/10.5281/zenodo.17426687 . We completed the rest of the checklist and are happy to move forward.
