### 2025-10-24 #8499 @cvoelt
@lucydot we noticed a small bug in figure 2, it was cut a tiny bit on the right side. We updated the figure and therefore also the paper. Thanks!
