### 2025-10-28 b1195d0
Bugfix Bibtex file
