+++ docs/paper.bib
@@ -24,7 +24,6 @@
 Publisher: Nature Publishing Group},
 	keywords = {Characterization and analytical techniques},
 	pages = {1--9},
-	file = {Full Text PDF:/home/constantin/CMDokumente/Studium/MA/Zotero/storage/3DUSJDB6/Starostin et al. - 2022 - Tracking perovskite crystallization via deep learn.pdf:application/pdf},
 }
 
 @article{starostin2022endtoend,
@@ -44,7 +43,6 @@ Publisher: Nature Publishing Group},
   booktitle = {Proceedings of the 27th ACM International Conference on Multimedia},
   series = {MM '19},
   year = {2019},
-  isbn = {978-1-4503-6889-6/19/10},
   location = {Nice, France},
   numpages = {4},
   url = {https://doi.org/10.1145/3343031.3350535},
@@ -102,7 +100,6 @@ url = {https://github.com/wkentaro/labelme}
 	year = {2015},
 	note = {Publisher: International Union of Crystallography},
 	pages = {917--926},
-	file = {jiang2015.pdf:/home/constantin/Downloads/jiang2015.pdf:application/pdf},
 }
 
 
@@ -121,7 +118,6 @@ url = {https://github.com/wkentaro/labelme}
 	year = {2017},
 	note = {Publisher: International Union of Crystallography},
 	pages = {959--966},
-	file = {Full Text PDF:/home/constantin/CMDokumente/Studium/MA/Zotero/storage/HBVW63HL/Filik et al. - 2017 - Processing two-dimensional X-ray diffraction and s.pdf:application/pdf},
 }
 
 
@@ -141,7 +137,6 @@ url = {https://github.com/wkentaro/labelme}
 	year = {2016},
 	note = {Publisher: International Union of Crystallography},
 	pages = {646--652},
-	file = {hammersley2016.pdf:/home/constantin/Downloads/hammersley2016.pdf:application/pdf},
 }
 
 
@@ -157,16 +152,13 @@ url = {https://github.com/wkentaro/labelme}
 	author = {Feidenhans'l, R.},
 	month = may,
 	year = {1989},
-	pages = {105--188},
-	file = {ScienceDirect Full Text PDF:/home/constantin/CMDokumente/Studium/MA/Zotero/storage/JLRA55MN/Feidenhans'l - 1989 - Surface structure determination by X-ray diffracti.pdf:application/pdf;ScienceDirect Snapshot:/home/constantin/CMDokumente/Studium/MA/Zotero/storage/DF7BEDN5/0167572989900022.html:text/html},
-}
+	pages = {105--188},}
 
 
 @incollection{banerjee2020grazing,
 	series = {World {Scientific} {Series} in {Nanoscience} and {Nanotechnology}},
 	title = {Grazing {Incidence} {X}-ray {Scattering} {Techniques} to {Study} {Growth} {Mechanism} of {Organic} {Films}},
 	volume = {Volume 21},
-	isbn = {978-981-12-3150-6},
 	url = {https://www.worldscientific.com/doi/abs/10.1142/9789811231513_0002},
 	number = {Volume 21},
 	urldate = {2024-07-23},
