### 2025-10-28 #8499 @cvoelt
Thank you @lucydot for checking again if ror links are allowed in the paper. We included it, since it is also in the example paper: https://joss.readthedocs.io/en/latest/example_paper.html . Regarding the error of the editorialbot, we believe it is caused by a malformed ISBN. We corrected it and hopefully the action will run without error.
