### 2025-11-04 #8499 @lucydot
> Thank you @lucydot for checking again if ror links are allowed in the paper. We included it, since it is also in the example paper: https://joss.readthedocs.io/en/latest/example_paper.html .

Yes, I had missed that!
